+++ Changelog.md
@@ -7,6 +7,12 @@ This project does _not_ adhere to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html) but contrary to it
 every new version is a new major version.
 
+## 106 - UNRELEASED
+
+### Added
+
+-   Nrfutil device `uart-write` support
+
 ## 105 - 2023-09-11
 
 ### Added